requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
idna==3.10
joblib==1.4.2
newsapi-python==0.2.7
nltk==3.9.1
python-dotenv==1.0.1
regex==2024.9.11
requests==2.32.3
soupsieve==2.6
textblob==0.18.0.post0
tqdm==4.66.5
urllib3==2.2.3
